minor codestyle fixes
